Move composite to sequence and the validators to validate
