+++ docs/source/tests/testing/binaries/sass.rst
@@ -10,13 +10,12 @@ The tests for the subpackages can be found in:
 
    sass/controlflow
    sass/instruction
-   sass/matchers
+   sass/operation
+   sass/sequence
 
 .. automodule:: tests.testing.binaries.sass.test_any
 .. automodule:: tests.testing.binaries.sass.test_atomic
 .. automodule:: tests.testing.binaries.sass.test_branch
-.. automodule:: tests.testing.binaries.sass.test_composite
-.. automodule:: tests.testing.binaries.sass.test_composite_impl
 .. automodule:: tests.testing.binaries.sass.test_extensibility
 .. automodule:: tests.testing.binaries.sass.test_instruction
 .. automodule:: tests.testing.binaries.sass.test_load
