--- docs/source/tests/testing/binaries/sass/matchers.rst
@@ -1,10 +1,0 @@
-Matchers
-========
-
-.. automodule:: tests.testing.binaries.sass.matchers.test_add_int128
-.. automodule:: tests.testing.binaries.sass.matchers.test_add_int32
-.. automodule:: tests.testing.binaries.sass.matchers.test_cas
-.. automodule:: tests.testing.binaries.sass.matchers.test_convert_fp_to_int
-.. automodule:: tests.testing.binaries.sass.matchers.test_convert_fp32_to_fp16
-.. automodule:: tests.testing.binaries.sass.matchers.test_convert_int_to_fp
-.. automodule:: tests.testing.binaries.sass.matchers.test_move32
