+++ docs/source/tests/testing/binaries/sass/operation.rst
@@ -1,0 +1,10 @@
+Operation
+=========
+
+.. automodule:: tests.testing.binaries.sass.operation.test_add_int128
+.. automodule:: tests.testing.binaries.sass.operation.test_add_int32
+.. automodule:: tests.testing.binaries.sass.operation.test_cas
+.. automodule:: tests.testing.binaries.sass.operation.test_convert_fp_to_int
+.. automodule:: tests.testing.binaries.sass.operation.test_convert_fp32_to_fp16
+.. automodule:: tests.testing.binaries.sass.operation.test_convert_int_to_fp
+.. automodule:: tests.testing.binaries.sass.operation.test_move32
