+++ docs/source/tests/testing/binaries/sass/sequence.rst
@@ -1,0 +1,4 @@
+Sequence
+========
+
+.. automodule:: tests.testing.binaries.sass.sequence.test_sequence
