+++ examples/kokkos/atomic/example_add_complex128.py
@@ -2,19 +2,19 @@ import re
 import sys
 import typing
 
-from reprospect.testing.binaries.sass.composite import (
-    unordered_interleaved_instructions_are,
-)
-from reprospect.testing.binaries.sass.composite_impl import UnorderedInSequenceMatcher
 from reprospect.testing.binaries.sass.instruction import (
+    Constant,
     InstructionMatch,
     OpcodeModsWithOperandsMatcher,
     PatternBuilder,
+    Register,
     RegisterMatcher,
 )
-from reprospect.testing.binaries.sass.instruction.constant import Constant
-from reprospect.testing.binaries.sass.instruction.register import Register
-from reprospect.testing.binaries.sass.matchers.cas import AtomicCASMatcher
+from reprospect.testing.binaries.sass.operation.cas import AtomicCASMatcher
+from reprospect.testing.binaries.sass.sequence import (
+    UnorderedInSequenceMatcher,
+    unordered_interleaved_instructions_are,
+)
 from reprospect.tools.binaries.sass import ControlFlow, Decoder
 
 from examples.kokkos.atomic import add, desul
