+++ examples/kokkos/atomic/example_add_int128.py
@@ -6,8 +6,8 @@ from reprospect.testing.binaries.sass.instruction import (
     InstructionMatch,
     RegisterMatcher,
 )
-from reprospect.testing.binaries.sass.matchers import add_int128
-from reprospect.testing.binaries.sass.matchers.cas import AtomicCASMatcher
+from reprospect.testing.binaries.sass.operation import add_int128
+from reprospect.testing.binaries.sass.operation.cas import AtomicCASMatcher
 from reprospect.tools.binaries.sass import ControlFlow, Decoder
 
 from examples.kokkos.atomic import add, desul
