+++ examples/kokkos/half/example_math.py
@@ -38,25 +38,23 @@ import typing
 import pytest
 
 from reprospect.testing import CMakeAwareTestCase
-from reprospect.testing.binaries.sass.composite import (
-    instruction_is,
-    instructions_are,
-    instructions_contain,
-)
-from reprospect.testing.binaries.sass.controlflow.block import BasicBlockMatcher
+from reprospect.testing.binaries.sass.controlflow import BasicBlockMatcher
 from reprospect.testing.binaries.sass.instruction import (
+    Fp16AddMatcher,
+    Fp16MinMaxMatcher,
     InstructionMatch,
     LoadGlobalMatcher,
     OpcodeModsWithOperandsMatcher,
+    Register,
     StoreGlobalMatcher,
 )
-from reprospect.testing.binaries.sass.instruction.half import (
-    Fp16AddMatcher,
-    Fp16MinMaxMatcher,
-)
-from reprospect.testing.binaries.sass.instruction.register import Register
-from reprospect.testing.binaries.sass.matchers.convert_fp32_to_fp16 import (
+from reprospect.testing.binaries.sass.operation.convert_fp32_to_fp16 import (
     ConvertFp32ToFp16,
+)
+from reprospect.testing.binaries.sass.sequence import (
+    instruction_is,
+    instructions_are,
+    instructions_contain,
 )
 from reprospect.tools.binaries import CuObjDump
 from reprospect.tools.binaries.sass import ControlFlow, Decoder, Instruction
