+++ pyproject.toml
@@ -35,7 +35,8 @@ packages = [
     "reprospect.testing.binaries.sass",
     "reprospect.testing.binaries.sass.controlflow",
     "reprospect.testing.binaries.sass.instruction",
-    "reprospect.testing.binaries.sass.matchers",
+    "reprospect.testing.binaries.sass.operation",
+    "reprospect.testing.binaries.sass.sequence",
     "reprospect.tools",
     "reprospect.tools.binaries",
     "reprospect.tools.binaries.sass",
