+++ reprospect/testing/binaries/sass/operation/add_int32.py
@@ -4,13 +4,11 @@ import typing
 from reprospect.testing.binaries.sass.instruction import (
     InstructionMatch,
     InstructionMatcher,
-    OpcodeModsWithOperandsMatcher,
-)
-from reprospect.testing.binaries.sass.instruction.integer import (
     IntAdd3Matcher,
     IntAddMatcher,
+    OpcodeModsWithOperandsMatcher,
+    Register,
 )
-from reprospect.testing.binaries.sass.instruction.register import Register
 from reprospect.tools.architecture import NVIDIAArch
 from reprospect.tools.binaries.sass.decode import Instruction
