+++ reprospect/testing/binaries/sass/sequence/__init__.py
@@ -1,0 +1,49 @@
+from .sequence import (
+    AllInSequenceMatcher,
+    AnyOfMatcher,
+    CountInSequenceMatcher,
+    InSequenceAtMatcher,
+    InSequenceMatcher,
+    OneOrMoreInSequenceMatcher,
+    OrderedInSequenceMatcher,
+    OrderedInterleavedInSequenceMatcher,
+    SequenceMatcher,
+    UnorderedInSequenceMatcher,
+    UnorderedInterleavedInSequenceMatcher,
+    ZeroOrMoreInSequenceMatcher,
+    any_of,
+    findall,
+    findunique,
+    instruction_count_is,
+    instruction_is,
+    instructions_are,
+    instructions_contain,
+    interleaved_instructions_are,
+    unordered_instructions_are,
+    unordered_interleaved_instructions_are,
+)
+
+__all__ = (
+    "AllInSequenceMatcher",
+    "AnyOfMatcher",
+    "CountInSequenceMatcher",
+    "InSequenceAtMatcher",
+    "InSequenceMatcher",
+    "OneOrMoreInSequenceMatcher",
+    "OrderedInSequenceMatcher",
+    "OrderedInterleavedInSequenceMatcher",
+    "SequenceMatcher",
+    "UnorderedInSequenceMatcher",
+    "UnorderedInterleavedInSequenceMatcher",
+    "ZeroOrMoreInSequenceMatcher",
+    "any_of",
+    "findall",
+    "findunique",
+    "instruction_count_is",
+    "instruction_is",
+    "instructions_are",
+    "instructions_contain",
+    "interleaved_instructions_are",
+    "unordered_instructions_are",
+    "unordered_interleaved_instructions_are",
+)
