+++ setup.py
@@ -19,8 +19,6 @@ def enable_mypyc(dist: Distribution) -> None:
     ext_modules = mypycify(
         [
             'reprospect/testing/features.py',
-            'reprospect/testing/binaries/sass/composite_impl.py',
-            'reprospect/testing/binaries/sass/composite.py',
             'reprospect/testing/binaries/sass/controlflow/block.py',
             'reprospect/testing/binaries/sass/instruction/address.py',
             'reprospect/testing/binaries/sass/instruction/atomic.py',
@@ -37,13 +35,15 @@ def enable_mypyc(dist: Distribution) -> None:
             'reprospect/testing/binaries/sass/instruction/pattern.py',
             'reprospect/testing/binaries/sass/instruction/register.py',
             'reprospect/testing/binaries/sass/instruction/store.py',
-            'reprospect/testing/binaries/sass/matchers/add_int128.py',
-            'reprospect/testing/binaries/sass/matchers/add_int32.py',
-            'reprospect/testing/binaries/sass/matchers/cas.py',
-            'reprospect/testing/binaries/sass/matchers/convert_fp_to_int.py',
-            'reprospect/testing/binaries/sass/matchers/convert_fp32_to_fp16.py',
-            'reprospect/testing/binaries/sass/matchers/convert_int_to_fp.py',
-            'reprospect/testing/binaries/sass/matchers/move32.py',
+            'reprospect/testing/binaries/sass/instruction/validate.py',
+            'reprospect/testing/binaries/sass/operation/add_int128.py',
+            'reprospect/testing/binaries/sass/operation/add_int32.py',
+            'reprospect/testing/binaries/sass/operation/cas.py',
+            'reprospect/testing/binaries/sass/operation/convert_fp_to_int.py',
+            'reprospect/testing/binaries/sass/operation/convert_fp32_to_fp16.py',
+            'reprospect/testing/binaries/sass/operation/convert_int_to_fp.py',
+            'reprospect/testing/binaries/sass/operation/move32.py',
+            'reprospect/testing/binaries/sass/sequence/sequence.py',
             'reprospect/tools/architecture.py',
             'reprospect/tools/binaries/cuobjdump.py',
             'reprospect/tools/binaries/demangle.py',
