+++ tests/testing/binaries/sass/CMakeLists.txt
@@ -1,12 +1,11 @@
 add_pytest(any)
 add_pytest(atomic)
 add_pytest(branch)
-add_pytest(composite_impl)
-add_pytest(composite)
 add_pytest(instruction)
 add_pytest(load)
 add_pytest(store)
 
 add_subdirectory(controlflow)
 add_subdirectory(instruction)
-add_subdirectory(matchers)
+add_subdirectory(operation)
+add_subdirectory(sequence)
