+++ tests/testing/binaries/sass/sequence/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_pytest(sequence)
